refactor: centralize file scanning util
