+++ chat/api_views.py
@@ -61,21 +61,9 @@ from .serializers import (
 from .services import criar_item_de_mensagem, sinalizar_mensagem
 from .tasks import exportar_historico_chat, gerar_resumo_chat
 from .throttles import FlagRateThrottle, UploadRateThrottle
+from .utils import _scan_file
 
 User = get_user_model()
-
-
-def _scan_file(path: str) -> bool:  # pragma: no cover - depends on external service
-    try:
-        import clamd  # type: ignore
-
-        cd = clamd.ClamdNetworkSocket()
-        result = cd.scan(path)
-        if result:
-            return any(status == "FOUND" for _, (status, _) in result.items())
-    except Exception:
-        return False
-    return False
 
 
 class ChatChannelCategoryViewSet(viewsets.ModelViewSet):
