+++ chat/tasks.py
@@ -30,18 +30,7 @@ from .models import (
 
 User = get_user_model()
 
-
-def _scan_file(path: str) -> bool:  # pragma: no cover - depends on external service
-    try:
-        import clamd  # type: ignore
-
-        cd = clamd.ClamdNetworkSocket()
-        result = cd.scan(path)
-        if result:
-            return any(status == "FOUND" for _, (status, _) in result.items())
-    except Exception:
-        return False
-    return False
+from .utils import _scan_file
 
 
 @shared_task
